+++ transaction/signer/src/types.rs
@@ -4,7 +4,6 @@
 //! wallet implementations.
 
 use mc_core::{
-    account::AccountId,
     keys::{RootSpendPublic, RootViewPrivate, TxOutPublic},
 };
 
@@ -150,6 +149,54 @@ pub struct TxSignResp {
 
     /// Mapping of real Tx public keys to key images
     pub txos: Vec<TxoSynced>,
+}
+
+
+/// Account ID object, derived from an [AccountKey] and used to identify
+/// individual accounts.
+#[derive(Clone, Eq, PartialEq, Hash)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+pub struct AccountId([u8; 32]);
+
+/// Display [AccountId] as a hex encoded string
+impl core::fmt::Display for AccountId {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        for v in self.0 {
+            write!(f, "{v:02X}")?;
+        }
+        Ok(())
+    }
+}
+
+impl core::fmt::Debug for AccountId {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(f, "AccountId(")?;
+        for v in self.0 {
+            write!(f, "{v:02X}")?;
+        }
+        write!(f, ")")
+    }
+}
+
+/// Access raw [AccountId] hash
+impl AsRef<[u8; 32]> for AccountId {
+    fn as_ref(&self) -> &[u8; 32] {
+        &self.0
+    }
+}
+
+/// Create [AccountId] object from raw hash
+impl From<[u8; 32]> for AccountId {
+    fn from(value: [u8; 32]) -> Self {
+        Self(value)
+    }
+}
+
+/// Create [AccountId] object from raw hash
+impl From<&[u8; 32]> for AccountId {
+    fn from(value: &[u8; 32]) -> Self {
+        Self(*value)
+    }
 }
 
 /// Public key hex encoding support for serde
